Change SMEWebify/WebErpMesv2/issues/506

+++ database/migrations/2021_09_13_183115_create_quotes_table.php
@@ -19,9 +19,9 @@ class CreateQuotesTable extends Migration
             $table->string('code');
 			$table->string('label');
 			$table->string('customer_reference')->nullable();
-			$table->integer('companies_id');
-			$table->integer('companies_contacts_id');
-			$table->integer('companies_addresses_id');
+			$table->unsignedBigInteger('companies_id');
+			$table->unsignedBigInteger('companies_contacts_id');
+			$table->unsignedBigInteger('companies_addresses_id');
 			$table->date('validity_date')->nullable();
 			$table->integer('statu')->default(1);
             #1 = Open
@@ -30,11 +30,11 @@ class CreateQuotesTable extends Migration
             #4 = Lost
             #5 = Closed
             #6 = Obsolete
-			$table->integer('user_id');
-			$table->integer('opportunities_id')->nullable();
-			$table->integer('accounting_payment_conditions_id');
-			$table->integer('accounting_payment_methods_id');
-			$table->integer('accounting_deliveries_id');
+			$table->unsignedBigInteger('user_id');
+			$table->unsignedBigInteger('opportunities_id')->nullable();
+			$table->unsignedBigInteger('accounting_payment_conditions_id');
+			$table->unsignedBigInteger('accounting_payment_methods_id');
+			$table->unsignedBigInteger('accounting_deliveries_id');
 			$table->text('comment')->nullable();
             $table->timestamps();
