+++ database/migrations/2021_12_04_152455_create_deliverys_table.php
@@ -17,9 +17,9 @@ class CreateDeliverysTable extends Migration
             $table->id();
             $table->string('code');
 			$table->string('label');
-			$table->integer('companies_id');
-			$table->integer('companies_contacts_id');
-			$table->integer('companies_addresses_id');
+			$table->unsignedBigInteger('companies_id');
+			$table->unsignedBigInteger('companies_contacts_id');
+			$table->unsignedBigInteger('companies_addresses_id');
 			$table->integer('statu')->default(1);
             #1 = In progress
             #2 = Sent
@@ -28,7 +28,7 @@ class CreateDeliverysTable extends Migration
             #2 = Not chargeable
             #3 = Partly invoiced
             #4 = Invoiced
-			$table->integer('user_id');
+			$table->unsignedBigInteger('user_id');
 			$table->text('comment')->nullable();
             $table->integer('order_id')->nullable();
             $table->timestamps();
